+++ tests/testthat/test-0_vimp_rfsrc_S4.R
@@ -96,7 +96,6 @@ familiar:::test_all_vimp_methods_parallel(
 )
 
 
-
 # Continuous outcome -----------------------------------------------------------
 data <- familiar:::test_create_good_data("continuous")
 
@@ -119,12 +118,18 @@ testthat::test_that(
   paste0(
     "The RFSRC random forest minimum depth method correctly ranks continuous data."),
   {
-  vimp_table <- suppressWarnings(
-    familiar:::get_vimp_table(familiar:::.vimp(vimp_object, data)))
-
-  testthat::expect_equal(
-    any(vimp_table[rank <= 2]$name %in% c("enrltot", "avginc", "calwpct")),
-    TRUE)
+    vimp_table <- suppressWarnings(
+      familiar:::get_vimp_table(familiar:::.vimp(vimp_object, data)))
+    
+    # Expect that the vimp table has at most six rows.
+    testthat::expect_lte(nrow(vimp_table), 6L)
+    
+    # Expect that the names are the same as that of the features.
+    testthat::expect_true(
+      all(vimp_table$name %in% familiar:::get_feature_columns(data)))
+    
+    # Feature 1 is most important.
+    testthat::expect_equal(vimp_table[rank == 1, ]$name, "feature_1")
   }
 )
 
@@ -149,9 +154,15 @@ testthat::test_that(
     vimp_table <- suppressWarnings(
       familiar:::get_vimp_table(familiar:::.vimp(vimp_object, data)))
     
-    testthat::expect_equal(
-      any(vimp_table[rank <= 2]$name %in% c("enrltot", "avginc", "calwpct")),
-      TRUE)
+    # Expect that the vimp table has at most six rows.
+    testthat::expect_lte(nrow(vimp_table), 6L)
+    
+    # Expect that the names are the same as that of the features.
+    testthat::expect_true(
+      all(vimp_table$name %in% familiar:::get_feature_columns(data)))
+    
+    # Feature 1 is most important.
+    testthat::expect_equal(vimp_table[rank == 1, ]$name, "feature_1")
   }
 )
 
@@ -177,40 +188,18 @@ testthat::test_that(
     vimp_table <- suppressWarnings(
       familiar:::get_vimp_table(familiar:::.vimp(vimp_object, data)))
     
-    testthat::expect_equal(
-      any(vimp_table[rank <= 2]$name %in% c("enrltot", "avginc", "calwpct")),
-      TRUE)
-  }
-)
-
-# Process dataset.
-vimp_object <- familiar:::prepare_vimp_object(
-  data = data,
-  vimp_method = "random_forest_rfsrc_variable_hunting",
-  vimp_method_parameter_list = list(
-    "n_tree" = 4,
-    "sample_size" = 0.50,
-    "m_try" = 0.3,
-    "node_size" = 5,
-    "tree_depth" = 5,
-    "fs_vh_fold" = 3,
-    "fs_vh_n_rep" = 3),
-  outcome_type = "continuous",
-  cluster_method = "none",
-  imputation_method = "simple")
-
-testthat::test_that(
-  paste0(
-    "The RFSRC random forest variable hunting method correctly ranks continuous data."),
-  {
-    vimp_table <- suppressWarnings(
-      familiar:::get_vimp_table(familiar:::.vimp(vimp_object, data)))
+    # Expect that the vimp table has at most six rows.
+    testthat::expect_lte(nrow(vimp_table), 6L)
     
-    testthat::expect_equal(
-      all(vimp_table[rank <= 2]$name %in% familiar:::get_feature_columns(data)),
-      TRUE)
+    # Expect that the names are the same as that of the features.
+    testthat::expect_true(
+      all(vimp_table$name %in% familiar:::get_feature_columns(data)))
+    
+    # Feature 1 is most important.
+    testthat::expect_equal(vimp_table[rank == 1, ]$name, "feature_1")
   }
 )
+
 
 
 # Binomial outcome -------------------------------------------------------------
@@ -237,10 +226,15 @@ testthat::test_that(
     vimp_table <- suppressWarnings(
       familiar:::get_vimp_table(familiar:::.vimp(vimp_object, data)))
     
-    testthat::expect_equal(
-      all(vimp_table[rank <= 2]$name %in% c(
-        "cell_shape_uniformity", "clump_thickness", "epithelial_cell_size", "bare_nuclei")),
-      TRUE)
+    # Expect that the vimp table has at most six rows.
+    testthat::expect_lte(nrow(vimp_table), 6L)
+    
+    # Expect that the names are the same as that of the features.
+    testthat::expect_true(
+      all(vimp_table$name %in% familiar:::get_feature_columns(data)))
+    
+    # Feature 1 is most important.
+    testthat::expect_equal(vimp_table[rank == 1, ]$name, "feature_1")
   }
 )
 
@@ -266,10 +260,15 @@ testthat::test_that(
     vimp_table <- suppressWarnings(
       familiar:::get_vimp_table(familiar:::.vimp(vimp_object, data)))
     
-    testthat::expect_equal(
-      all(vimp_table[rank <= 2]$name %in% c(
-        "cell_shape_uniformity", "clump_thickness", "epithelial_cell_size", "bare_nuclei")),
-      TRUE)
+    # Expect that the vimp table has at most six rows.
+    testthat::expect_lte(nrow(vimp_table), 6L)
+    
+    # Expect that the names are the same as that of the features.
+    testthat::expect_true(
+      all(vimp_table$name %in% familiar:::get_feature_columns(data)))
+    
+    # Feature 1 is most important.
+    testthat::expect_equal(vimp_table[rank == 1, ]$name, "feature_1")
   }
 )
 
@@ -294,41 +293,17 @@ testthat::test_that(
     vimp_table <- suppressWarnings(
       familiar:::get_vimp_table(familiar:::.vimp(vimp_object, data)))
     
-    testthat::expect_equal(
-      all(vimp_table[rank <= 2]$name %in% familiar:::get_feature_columns(data)),
-      TRUE)
-  }
-)
-
-# Process dataset.
-vimp_object <- familiar:::prepare_vimp_object(
-  data = data,
-  vimp_method = "random_forest_rfsrc_variable_hunting",
-  vimp_method_parameter_list = list(
-    "n_tree" = 4,
-    "sample_size" = 0.50,
-    "m_try" = 0.3,
-    "node_size" = 5,
-    "tree_depth" = 5,
-    "fs_vh_fold" = 3,
-    "fs_vh_n_rep" = 3),
-  outcome_type = "binomial",
-  cluster_method = "none",
-  imputation_method = "simple")
-
-testthat::test_that(
-  paste0(
-    "The RFSRC random forest variable hunting method correctly ranks binomial data."),
-  {
-    vimp_table <- suppressWarnings(
-      familiar:::get_vimp_table(familiar:::.vimp(vimp_object, data)))
+    # Expect that the vimp table has at most six rows.
+    testthat::expect_lte(nrow(vimp_table), 6L)
     
-    testthat::expect_equal(
-      all(vimp_table[rank <= 2]$name %in% familiar:::get_feature_columns(data)),
-      TRUE)
+    # Expect that the names are the same as that of the features.
+    testthat::expect_true(
+      all(vimp_table$name %in% familiar:::get_feature_columns(data)))
+    
+    # Feature 1 is most important.
+    testthat::expect_equal(vimp_table[rank == 1, ]$name, "feature_1")
   }
 )
-
 
 
 # Multinomial outcome ----------------------------------------------------------
@@ -355,9 +330,15 @@ testthat::test_that(
     vimp_table <- suppressWarnings(
       familiar:::get_vimp_table(familiar:::.vimp(vimp_object, data)))
     
-    testthat::expect_equal(
-      all(vimp_table[rank <= 2]$name %in% c("Petal_Length", "Petal_Width")),
-      TRUE)
+    # Expect that the vimp table has at most six rows.
+    testthat::expect_lte(nrow(vimp_table), 6L)
+    
+    # Expect that the names are the same as that of the features.
+    testthat::expect_true(
+      all(vimp_table$name %in% familiar:::get_feature_columns(data)))
+    
+    # Feature 1 is most important.
+    testthat::expect_equal(vimp_table[rank == 1, ]$name, "feature_1")
   }
 )
 
@@ -382,9 +363,15 @@ testthat::test_that(
     vimp_table <- suppressWarnings(
       familiar:::get_vimp_table(familiar:::.vimp(vimp_object, data)))
     
-    testthat::expect_equal(
-      all(vimp_table[rank <= 2]$name %in% c("Petal_Length", "Petal_Width")),
-      TRUE)
+    # Expect that the vimp table has at most six rows.
+    testthat::expect_lte(nrow(vimp_table), 6L)
+    
+    # Expect that the names are the same as that of the features.
+    testthat::expect_true(
+      all(vimp_table$name %in% familiar:::get_feature_columns(data)))
+    
+    # Feature 1 is most important.
+    testthat::expect_equal(vimp_table[rank == 1, ]$name, "feature_1")
   }
 )
 
@@ -409,41 +396,17 @@ testthat::test_that(
     vimp_table <- suppressWarnings(
       familiar:::get_vimp_table(familiar:::.vimp(vimp_object, data)))
     
-    testthat::expect_equal(
-      all(vimp_table[rank <= 2]$name %in% familiar:::get_feature_columns(data)),
-      TRUE)
+    # Expect that the vimp table has at most six rows.
+    testthat::expect_lte(nrow(vimp_table), 6L)
+    
+    # Expect that the names are the same as that of the features.
+    testthat::expect_true(
+      all(vimp_table$name %in% familiar:::get_feature_columns(data)))
+    
+    # Feature 1 is most important.
+    testthat::expect_equal(vimp_table[rank == 1, ]$name, "feature_1")
   }
 )
-
-# Process dataset.
-vimp_object <- familiar:::prepare_vimp_object(
-  data = data,
-  vimp_method = "random_forest_rfsrc_variable_hunting",
-  vimp_method_parameter_list = list(
-    "n_tree" = 4,
-    "sample_size" = 0.50,
-    "m_try" = 0.3,
-    "node_size" = 5,
-    "tree_depth" = 5,
-    "fs_vh_fold" = 3,
-    "fs_vh_n_rep" = 3),
-  outcome_type = "multinomial",
-  cluster_method = "none",
-  imputation_method = "simple")
-
-testthat::test_that(
-  paste0(
-    "The RFSRC random forest variable hunting method correctly ranks multinomial outcome data."),
-  {
-    vimp_table <- suppressWarnings(
-      familiar:::get_vimp_table(familiar:::.vimp(vimp_object, data)))
-
-    testthat::expect_equal(
-      all(vimp_table[rank <= 2]$name %in% familiar:::get_feature_columns(data)),
-      TRUE)
-  }
-)
-
 
 
 # Survival outcome -------------------------------------------------------------
@@ -470,9 +433,15 @@ testthat::test_that(
     vimp_table <- suppressWarnings(
       familiar:::get_vimp_table(familiar:::.vimp(vimp_object, data)))
     
-    testthat::expect_equal(
-      all(vimp_table[rank <= 2]$name %in% c("nodes", "rx", "adhere")), 
-      TRUE)
+    # Expect that the vimp table has at most six rows.
+    testthat::expect_lte(nrow(vimp_table), 6L)
+    
+    # Expect that the names are the same as that of the features.
+    testthat::expect_true(
+      all(vimp_table$name %in% familiar:::get_feature_columns(data)))
+    
+    # Feature 1 is most important.
+    testthat::expect_equal(vimp_table[rank == 1, ]$name, "feature_1")
   }
 )
 
@@ -496,9 +465,15 @@ testthat::test_that(
       vimp_table <- suppressWarnings(
         familiar:::get_vimp_table(familiar:::.vimp(vimp_object, data)))
       
-      testthat::expect_equal(
-        all(vimp_table[rank <= 2]$name %in% c("nodes", "rx")),
-        TRUE)
+      # Expect that the vimp table has at most six rows.
+      testthat::expect_lte(nrow(vimp_table), 6L)
+      
+      # Expect that the names are the same as that of the features.
+      testthat::expect_true(
+        all(vimp_table$name %in% familiar:::get_feature_columns(data)))
+      
+      # Feature 1 is most important.
+      testthat::expect_equal(vimp_table[rank == 1, ]$name, "feature_1")
     }
 )
 
@@ -524,40 +499,18 @@ testthat::test_that(
     vimp_table <- suppressWarnings(
       familiar:::get_vimp_table(familiar:::.vimp(vimp_object, data)))
     
-    testthat::expect_equal(
-      all(vimp_table[rank <= 2]$name %in% c("nodes", "rx")), 
-      TRUE)
-  }
-)
-
-# Process dataset.
-vimp_object <- familiar:::prepare_vimp_object(
-  data = data,
-  vimp_method = "random_forest_rfsrc_variable_hunting",
-  vimp_method_parameter_list = list(
-    "n_tree" = 4,
-    "sample_size" = 0.50,
-    "m_try" = 0.3,
-    "node_size" = 5,
-    "tree_depth" = 5,
-    "fs_vh_fold" = 3,
-    "fs_vh_n_rep" = 3),
-  outcome_type = "survival",
-  cluster_method = "none",
-  imputation_method = "simple")
-
-testthat::test_that(
-  paste0(
-    "The RFSRC random forest variable hunting method correctly ranks survival outcome data."),
-  {
-    vimp_table <- suppressWarnings(
-      familiar:::get_vimp_table(familiar:::.vimp(vimp_object, data)))
+    # Expect that the vimp table has at most six rows.
+    testthat::expect_lte(nrow(vimp_table), 6L)
     
-    testthat::expect_equal(
-      all(vimp_table[rank <= 2]$name %in% familiar:::get_feature_columns(data)),
-      TRUE)
+    # Expect that the names are the same as that of the features.
+    testthat::expect_true(
+      all(vimp_table$name %in% familiar:::get_feature_columns(data)))
+    
+    # Feature 1 is most important.
+    testthat::expect_equal(vimp_table[rank == 1, ]$name, "feature_1")
   }
 )
+
 
 testthat::skip("Skip hyperparameter optimisation, unless manual.")